当前STM32内部Flash通过测试，可以使用

+++ HARDWARE/STMFLASH/stmflash.h
@@ -23,7 +23,7 @@
 //使用一半空间作为系统文件资源
 #define STM32_FLASH_FLLESYS_START_BASE   (0x08000000 + STM32Flash_ERASE_GRAN*128)
 
-uint8_t STMFLASH_ReadHalfWord(u32 faddr);		  //读出半字  
+uint16_t STMFLASH_ReadHalfWord(u32 faddr);		  //读出半字  
 void STMFLASH_WriteLenByte(u32 WriteAddr,u32 DataToWrite,u16 Len);	//指定地址开始写入指定长度的数据
 u32 STMFLASH_ReadLenByte(u32 ReadAddr,u16 Len);						//指定地址开始读取指定长度数据
 void STMFLASH_Write(u32 WriteAddr,u16 *pBuffer,u16 NumToWrite);		//从指定地址开始写入指定长度的数据
